Refactor a bit, make codacy a bit more happy + make sure to get harvesing behaviour from entityData. Next step is get from ini file

+++ src/main/java/com/fundynamic/d2tm/game/map/Cell.java
@@ -163,11 +163,11 @@ public class Cell {
         return terrain instanceof Harvestable;
     }
 
-    public int harvest(int amount) {
+    public float harvest(float amount) {
         if (isHarvestable()) {
             return ((Harvestable) terrain).harvest(amount);
         }
-        return 0;
+        return 0f;
     }
 
     public float distance(Entity entity) {
