+++ src/test/java/com/fundynamic/d2tm/game/terrain/impl/HarvestableTerrain.java
@@ -16,7 +16,7 @@ public class HarvestableTerrain extends EmptyTerrain implements Harvestable {
     }
 
     @Override
-    public int harvest(int amount) {
+    public float harvest(float amount) {
         return this.amount -= amount;
     }
 }
